nextjs-project-setup-main/youtube_summarizer/requirements.txt
# Core dependencies
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.0.0
youtube-transcript-api>=0.6.0

# Text processing
nltk>=3.8.1
spacy>=3.7.0
sentence-transformers>=2.2.2

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# File handling and utilities
pathlib
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF generation (optional)
reportlab>=4.0.0
fpdf2>=2.7.0

# Progress bars and UI enhancements
tqdm>=4.66.0

# Audio processing (fallback for Whisper)
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0

# Additional utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
